+++ tests/Command/CodeFixerCommandTest.php
@@ -76,7 +76,7 @@ EOT;
 
         // Verify various parts of the output.
         $output = $commandTester->getDisplay();
-        $this->assertMatchesRegularExpression('/F\.* 8\.* \/ 8 \(100%\)/', $output);                   // Progress.
+        $this->assertMatchesRegularExpression('/F\.* 10\.* \/ 10 \(100%\)/', $output);                // Progress.
         $this->assertMatchesRegularExpression('/\/fixable.php/', $output);                            // File.
         $this->assertMatchesRegularExpression('/A TOTAL OF 1 ERROR WERE FIXED IN 1 FILE/', $output);  // Summary.
         $this->assertMatchesRegularExpression('/Time:.*Memory:/', $output);                           // Time.